[Sylius] Extract CoreBundle's Kernel once again
