+++ app/AppKernel.php
@@ -9,9 +9,7 @@
  * file that was distributed with this source code.
  */
 
-use PSS\SymfonyMockerContainer\DependencyInjection\MockerContainer;
-use Symfony\Component\Config\Loader\LoaderInterface;
-use Symfony\Component\HttpKernel\Kernel;
+use Sylius\Bundle\CoreBundle\Application\Kernel;
 
 /**
  * @author Paweł Jędrzejewski <pawel@sylius.org>
@@ -19,165 +17,15 @@ use Symfony\Component\HttpKernel\Kernel;
  */
 class AppKernel extends Kernel
 {
-    const VERSION = '0.19.0-dev';
-    const VERSION_ID = '00190';
-    const MAJOR_VERSION = '0';
-    const MINOR_VERSION = '19';
-    const RELEASE_VERSION = '0';
-    const EXTRA_VERSION = 'DEV';
-
     /**
      * {@inheritdoc}
      */
     public function registerBundles()
     {
         $bundles = [
-            new \Sylius\Bundle\InstallerBundle\SyliusInstallerBundle(),
-            new \Sylius\Bundle\OrderBundle\SyliusOrderBundle(),
-            new \Sylius\Bundle\MoneyBundle\SyliusMoneyBundle(),
-            new \Sylius\Bundle\CurrencyBundle\SyliusCurrencyBundle(),
-            new \Sylius\Bundle\ContactBundle\SyliusContactBundle(),
-            new \Sylius\Bundle\LocaleBundle\SyliusLocaleBundle(),
-            new \Sylius\Bundle\SettingsBundle\SyliusSettingsBundle(),
-            new \Sylius\Bundle\CartBundle\SyliusCartBundle(),
-            new \Sylius\Bundle\ProductBundle\SyliusProductBundle(),
-            new \Sylius\Bundle\ArchetypeBundle\SyliusArchetypeBundle(),
-            new \Sylius\Bundle\ChannelBundle\SyliusChannelBundle(),
-            new \Sylius\Bundle\VariationBundle\SyliusVariationBundle(),
-            new \Sylius\Bundle\AttributeBundle\SyliusAttributeBundle(),
-            new \Sylius\Bundle\TaxationBundle\SyliusTaxationBundle(),
-            new \Sylius\Bundle\ShippingBundle\SyliusShippingBundle(),
-            new \Sylius\Bundle\PaymentBundle\SyliusPaymentBundle(),
-            new \Sylius\Bundle\MailerBundle\SyliusMailerBundle(),
-            new \Sylius\Bundle\ReportBundle\SyliusReportBundle(),
-            new \Sylius\Bundle\PromotionBundle\SyliusPromotionBundle(),
-            new \Sylius\Bundle\AddressingBundle\SyliusAddressingBundle(),
-            new \Sylius\Bundle\InventoryBundle\SyliusInventoryBundle(),
-            new \Sylius\Bundle\TaxonomyBundle\SyliusTaxonomyBundle(),
-            new \Sylius\Bundle\FlowBundle\SyliusFlowBundle(),
-            new \Sylius\Bundle\PricingBundle\SyliusPricingBundle(),
-            new \Sylius\Bundle\SequenceBundle\SyliusSequenceBundle(),
-            new \Sylius\Bundle\ContentBundle\SyliusContentBundle(),
-            new \Sylius\Bundle\SearchBundle\SyliusSearchBundle(),
-            new \Sylius\Bundle\RbacBundle\SyliusRbacBundle(),
-            new \Sylius\Bundle\UserBundle\SyliusUserBundle(),
-            new \Sylius\Bundle\UiBundle\SyliusUiBundle(),
-            new \Sylius\Bundle\AdminBundle\SyliusAdminBundle(),
-            new \Sylius\Bundle\ShopBundle\SyliusShopBundle(),
-            new \Sylius\Bundle\MetadataBundle\SyliusMetadataBundle(),
-            new \Sylius\Bundle\AssociationBundle\SyliusAssociationBundle(),
-            new \Sylius\Bundle\ReviewBundle\SyliusReviewBundle(),
-            new \Sylius\Bundle\CoreBundle\SyliusCoreBundle(),
-            new \Sylius\Bundle\WebBundle\SyliusWebBundle(),
-            new \Sylius\Bundle\ResourceBundle\SyliusResourceBundle(),
-            new \Sylius\Bundle\GridBundle\SyliusGridBundle(),
-            new \winzou\Bundle\StateMachineBundle\winzouStateMachineBundle(),
-            new \Sylius\Bundle\ApiBundle\SyliusApiBundle(),
 
-            new \Sonata\BlockBundle\SonataBlockBundle(),
-            new \Symfony\Cmf\Bundle\CoreBundle\CmfCoreBundle(),
-            new \Symfony\Cmf\Bundle\BlockBundle\CmfBlockBundle(),
-            new \Symfony\Cmf\Bundle\ContentBundle\CmfContentBundle(),
-            new \Symfony\Cmf\Bundle\RoutingBundle\CmfRoutingBundle(),
-            new \Symfony\Cmf\Bundle\MenuBundle\CmfMenuBundle(),
-            new \Symfony\Cmf\Bundle\CreateBundle\CmfCreateBundle(),
-            new \Symfony\Cmf\Bundle\MediaBundle\CmfMediaBundle(),
-
-            new \Doctrine\Bundle\DoctrineBundle\DoctrineBundle(),
-            new \Doctrine\Bundle\DoctrineCacheBundle\DoctrineCacheBundle(),
-            new \Doctrine\Bundle\PHPCRBundle\DoctrinePHPCRBundle(),
-            new \Symfony\Bundle\AsseticBundle\AsseticBundle(),
-            new \Symfony\Bundle\FrameworkBundle\FrameworkBundle(),
-            new \Symfony\Bundle\MonologBundle\MonologBundle(),
-            new \Symfony\Bundle\SecurityBundle\SecurityBundle(),
-            new \Symfony\Bundle\SwiftmailerBundle\SwiftmailerBundle(),
-            new \Symfony\Bundle\TwigBundle\TwigBundle(),
-
-            new \Bazinga\Bundle\HateoasBundle\BazingaHateoasBundle(),
-            new \FOS\OAuthServerBundle\FOSOAuthServerBundle(),
-            new \FOS\RestBundle\FOSRestBundle(),
-
-            new \FOS\ElasticaBundle\FOSElasticaBundle(),
-            new \Knp\Bundle\GaufretteBundle\KnpGaufretteBundle(),
-            new \Knp\Bundle\MenuBundle\KnpMenuBundle(),
-            new \Knp\Bundle\SnappyBundle\KnpSnappyBundle(),
-            new \Liip\ImagineBundle\LiipImagineBundle(),
-            new \Payum\Bundle\PayumBundle\PayumBundle(),
-            new \JMS\SerializerBundle\JMSSerializerBundle(),
-            new \JMS\TranslationBundle\JMSTranslationBundle(),
-            new \Stof\DoctrineExtensionsBundle\StofDoctrineExtensionsBundle(),
-            new \WhiteOctober\PagerfantaBundle\WhiteOctoberPagerfantaBundle(),
-
-            new \Doctrine\Bundle\MigrationsBundle\DoctrineMigrationsBundle(),
-            new \Doctrine\Bundle\FixturesBundle\DoctrineFixturesBundle(),
-            new \Sylius\Bundle\FixturesBundle\SyliusFixturesBundle(),
-            new \Sylius\Bundle\PayumBundle\SyliusPayumBundle(), // must be added after PayumBundle.
-            new \Sylius\Bundle\ThemeBundle\SyliusThemeBundle(), // must be added after FrameworkBundle
         ];
 
-        if (in_array($this->environment, ['dev', 'test'], true)) {
-            $bundles[] = new \Symfony\Bundle\DebugBundle\DebugBundle();
-            $bundles[] = new \Symfony\Bundle\WebProfilerBundle\WebProfilerBundle();
-        }
-
-        return $bundles;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function getContainerBaseClass()
-    {
-        if ('test' === $this->environment || 'test_cached' === $this->environment) {
-            return MockerContainer::class;
-        }
-
-        return parent::getContainerBaseClass();
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function registerContainerConfiguration(LoaderInterface $loader)
-    {
-        $rootDir = $this->getRootDir();
-
-        $loader->load($rootDir.'/config/config_'.$this->environment.'.yml');
-
-        if (is_file($file = $rootDir.'/config/config_'.$this->environment.'.local.yml')) {
-            $loader->load($file);
-        }
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getCacheDir()
-    {
-        if ($this->isVagrantEnvironment()) {
-            return '/dev/shm/sylius/cache/'.$this->environment;
-        }
-
-        return parent::getCacheDir();
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getLogDir()
-    {
-        if ($this->isVagrantEnvironment()) {
-            return '/dev/shm/sylius/logs';
-        }
-
-        return parent::getLogDir();
-    }
-
-    /**
-     * @return bool
-     */
-    protected function isVagrantEnvironment()
-    {
-        return (getenv('HOME') === '/home/vagrant' || getenv('VAGRANT') === 'VAGRANT') && is_dir('/dev/shm');
+        return array_merge(parent::registerBundles(), $bundles);
     }
 }
