+++ src/Sylius/Bundle/CoreBundle/DataCollector/SyliusDataCollector.php
@@ -11,6 +11,7 @@
 
 namespace Sylius\Bundle\CoreBundle\DataCollector;
 
+use Sylius\Bundle\CoreBundle\Application\Kernel;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\DataCollector\DataCollector;
@@ -22,7 +23,7 @@ class SyliusDataCollector extends DataCollector
      */
     public function collect(Request $request, Response $response, \Exception $exception = null)
     {
-        $this->data['version'] = \AppKernel::VERSION;
+        $this->data['version'] = Kernel::VERSION;
     }
 
     /**
